# src/main/resources/openapi/swagger-config.yaml
...
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  "/api/users/update/{id}":
    patch:
      tags:
        - user-controller
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UserDTO"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserDTO"
...
  "/api/users/get/{id}":
    get:
      tags:
        - user-controller
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserDTO"
...
  "/api/users/delete/{id}":
    delete:
      tags:
        - user-controller
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
...
components:
  schemas:
    TaskDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 1
        name:
          type: string
          example: This is the name of task
        description:
          type: string
          example: This is the description of task
        created_at:
          type: string
          format: date-time
          readOnly: true
        task_state_id:
          type: integer
          format: int64
          readOnly: true
          example: 1
      example:
        - id: 27
          name: task1
          description: description1
          created_at: '2023-10-24T17:26:05.902537Z'
          task_state_id: 146
    TaskStateDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 1
        name:
          type: string
          example: This is the name of task state
        previous_task_state_id:
          type: integer
          format: int64
          example: 1
        next_task_state_id:
          type: integer
          format: int64
          readOnly: true
          example: 1
        created_at:
          type: string
          format: date-time
          readOnly: true
        project_id:
          type: array
          readOnly: true
          example:
            - id: 27
              name: task1
              description: description1
              created_at: '2023-10-24T17:26:05.902537Z'
              task_state_id: 146
          items:
            "$ref": "#/components/schemas/TaskDTO"
      example:
        - id: 145
          name: In progress
          previous_task_state_id: 144
          next_task_state_id: 146
          created_at: '2023-10-24T17:16:43.498700Z'
          project_id: 57
          task_entities: []
    ProjectDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 1
        name:
          type: string
          example: This is the name of project
        created_at:
          type: string
          format: date-time
          readOnly: true
        user_id:
          type: integer
          format: int64
          readOnly: true
        task_state_entities:
          type: array
          readOnly: true
          example:
            - id: 145
              name: In progress
              previous_task_state_id: 144
              next_task_state_id: 146
              created_at: '2023-10-24T17:16:43.498700Z'
              project_id: 57
              task_entities: []
          items:
            "$ref": "#/components/schemas/TaskStateDTO"
...
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        username:
          type: string
        password:
          type: string
          writeOnly: true
        email:
          type: string
        projects:
          type: array
          readOnly: true
          items:
            "$ref": "#/components/schemas/ProjectDTO"